feat: add user entity
