--- src/user/interfaces/user.interface.ts
@@ -1,10 +1,0 @@
-export interface User {
-  id: string;
-  login: string;
-  password: string;
-  version: number;
-  createdAt: number;
-  updatedAt: number;
-}
-
-export type UserWithoutPassword = Omit<User, 'password'>;
