+++ src/user/user.entity.ts
@@ -1,0 +1,24 @@
+import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm';
+
+@Entity()
+export class User {
+  @PrimaryGeneratedColumn('uuid')
+  id: string;
+
+  @Column()
+  login: string;
+
+  @Column()
+  password: string;
+
+  @Column()
+  version: number;
+
+  @Column('bigint')
+  createdAt: number;
+
+  @Column('bigint')
+  updatedAt: number;
+}
+
+export type UserWithoutPassword = Omit<User, 'password'>;
